+++ src/main/java/kr/co/pawong/pwbe/adoption/application/service/ApiRequestService.java
@@ -10,16 +10,16 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import kr.co.pawong.pwbe.adoption.application.port.out.dto.AdoptionCreate;
+import kr.co.pawong.pwbe.adoption.application.port.in.dto.AdoptionCreate;
 import kr.co.pawong.pwbe.adoption.enums.NeuterYn;
 import kr.co.pawong.pwbe.adoption.enums.ProcessState;
 import kr.co.pawong.pwbe.adoption.enums.SexCd;
 import kr.co.pawong.pwbe.adoption.enums.UpKindCd;
 import kr.co.pawong.pwbe.adoption.enums.UpKindNm;
-import kr.co.pawong.pwbe.adoption.application.port.out.dto.AdoptionApi;
-import kr.co.pawong.pwbe.adoption.application.port.out.dto.AdoptionApi.Body;
-import kr.co.pawong.pwbe.adoption.application.port.out.dto.AdoptionApi.Items;
-import kr.co.pawong.pwbe.adoption.application.port.out.dto.AdoptionApi.Response;
+import kr.co.pawong.pwbe.adoption.application.service.dto.AdoptionApi;
+import kr.co.pawong.pwbe.adoption.application.service.dto.AdoptionApi.Body;
+import kr.co.pawong.pwbe.adoption.application.service.dto.AdoptionApi.Items;
+import kr.co.pawong.pwbe.adoption.application.service.dto.AdoptionApi.Response;
 import kr.co.pawong.pwbe.adoption.application.port.in.UpdateDataUseCase;
 import kr.co.pawong.pwbe.adoption.application.port.in.ApiRequestUseCase;
 import lombok.RequiredArgsConstructor;
