+++ src/main/java/kr/co/pawong/pwbe/shelter/application/port/in/QueryShelterDataUseCase.java
@@ -1,0 +1,11 @@
+package kr.co.pawong.pwbe.shelter.application.port.in;
+
+import kr.co.pawong.pwbe.shelter.application.port.in.dto.ShelterDetailDto;
+import kr.co.pawong.pwbe.shelter.application.port.in.dto.ShelterInfoDto;
+
+public interface QueryShelterDataUseCase {
+
+    ShelterInfoDto shelterInfo(String careRegNo);
+
+    ShelterDetailDto shelterDetail(String careRegNo);
+}
