+++ src/main/java/kr/co/pawong/pwbe/shelter/application/port/in/UpdateShelterDataUseCase.java
@@ -1,0 +1,11 @@
+package kr.co.pawong.pwbe.shelter.application.port.in;
+
+import kr.co.pawong.pwbe.shelter.application.port.in.dto.ShelterCreate;
+import java.util.List;
+
+
+public interface UpdateShelterDataUseCase {
+    // ShelterCreate -> Shelter
+    void saveShelters(List<ShelterCreate> shelterCreates);
+
+}
