--- src/main/java/kr/co/pawong/pwbe/shelter/application/service/port/ShelterQueryRepository.java
@@ -1,10 +1,0 @@
-package kr.co.pawong.pwbe.shelter.application.service.port;
-
-import kr.co.pawong.pwbe.shelter.application.domain.Shelter;
-import kr.co.pawong.pwbe.shelter.presentation.controller.dto.ShelterInfoDto;
-
-
-public interface ShelterQueryRepository {
-
-    Shelter findByCareRegNoOrThrow(String careRegNo);
-}
